pyproject.toml
[project]
name = "website-about-novaya-lyada"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "django-allauth>=65.10.0",
    "django-cleanup>=9.0.0",
    "django-password-eye>=1.1.4",
    # tests
    "pytest-django>=4.9.0",
    "django-sslserver-v2>=1.0",
    "python-dotenv>=1.0.1",
    "pillow>=11.3.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "requests>=2.32.4",
    "social-auth-app-django>=5.5.1",
    "django-imagekit>=5.0.0",
    "celery>=5.4.0",
    "redis>=5.0.0",
    "python-slugify>=8.0.0",
    "psycopg2-binary>=2.9.0",
    "gunicorn>=21.2.0",
]

[dependency-groups]
dev = [
    "pyopenssl>=25.1.0",
]

[tool.setuptools.packages.find]
include = ["website_about_novaya_lyada*"]
exclude = ["book_images*", "book_extracted*", "scripts*", ".github*", "nl_website*"]


